refactor: break transports into a separate crate
