+++ Cargo.toml
@@ -16,6 +16,7 @@ Complete Ethereum library and wallet implementation in Rust.
 members = [
     "ethers-addressbook",
     "ethers-contract",
+    "ethers-transports",
     "ethers-providers",
     "ethers-signers",
     "ethers-core",
@@ -28,6 +29,7 @@ members = [
 default-members = [
     "ethers-addressbook",
     "ethers-contract",
+    "ethers-transports",
     "ethers-providers",
     "ethers-signers",
     "ethers-core",
@@ -85,6 +87,7 @@ solc-sha2-asm = ["ethers-solc/asm"]
 ethers-addressbook = { version = "^0.1.0", default-features = false, path = "./ethers-addressbook" }
 ethers-contract = { version = "^0.6.0", default-features = false, path = "./ethers-contract" }
 ethers-core = { version = "^0.6.0", default-features = false, path = "./ethers-core" }
+ethers-transports = { version = "^0.6.0", default-features = false, path = "./ethers-transports" }
 ethers-providers = { version = "^0.6.0", default-features = false, path = "./ethers-providers" }
 ethers-signers = { version = "^0.6.0", default-features = false, path = "./ethers-signers" }
 ethers-middleware = { version = "^0.6.0", default-features = false, path = "./ethers-middleware" }
