+++ ethers-transports/src/rw.rs
@@ -1,7 +1,7 @@
 //! A [JsonRpcClient] implementation that serves as a wrapper around two different [JsonRpcClient]
 //! and uses a dedicated client for read and the other for write operations
 
-use crate::{provider::ProviderError, JsonRpcClient};
+use crate::{ProviderError, JsonRpcClient};
 
 use async_trait::async_trait;
